Automatically record to USB drive when present
If a volume labelled "cp" is mounted, the set-thermal-recorder-output
service is started. This reconfigures thermal-recorder to use the
external volume and restarts it. Any CPTV files in /var/spool/CPTV are
also moved to the volume.

When/if the drive is removed, thermal-recorder is reconfigured to
write to /var/spool/CPTV again.