Use telemetry data to identify FFC instead of guessing
The "last FFC time" field in the camera's telemetry data is now used
to decide whether readings could be affected by a recent
recalibration, instead of guessing by looking for large changes in the
temperature readings.

The `nonzero-max-percent` option has been removed from the
configuration file.

Many tests and test helpers needed restructuring to account for the
new way that FFC events are handled.

testframemaker.go was renamed to testframemaker_test.go to ensure it
is only built for test runs and cannot be included in production builds.