Work around camera initialisation issue
It turns out when GPIO23 is low the camera's Vin->GND voltage was only
dropping to 2.5V instead of 0V. It seems the camera is still getting
some kind of power via other pins. This seems to sometimes make the
camera fail to reset properly (more likely in some devices than
others).

Uninstalling the SPI driver disables more of the camera's pins and
allows the voltage to drop to 1.2V, allowing the camera to reset
reliably.

This also seems to have fixed the I2C initialisation issues that we
typically see on the first start of leptond after boot.