Recording will restart if bucket refills while recording in progress
The previous implementation would never allow a recording to restart
once it had been throttled, even if the bucket refilled during the
recording. This could result in longer than expected amounts of
throttling.

This change also allowed for significant simplification of the
throttling code, reducing the amount of state stored.